\name{AROC-package}
\alias{AROC-package}
\alias{AROC}
\docType{package}
\title{
\packageTitle{AROC}
}
\description{
Estimates the covariate-adjusted Receiver Operating Characteristic (AROC) curve and pooled (unadjusted) ROC curve by different methods.
}
\details{
\tabular{ll}{
Package: \tab AROC\cr
Type: \tab Package\cr
Version: \tab 1.0-4\cr
Date: \tab 2022-02-18\cr
License: \tab GPL\cr
}
}
\author{
Maria Xose Rodriguez-Alvarez and Vanda Inacio.

Maintainer: Maria Xose Rodriguez-Alvarez <mxrodriguez@uvigo.es>
}
\references{

Janes, H., and Pepe, M.S. (2009). Adjusting for covariate effects on classification accuracy using the covariate-adjusted receiver operating characteristic curve. Biometrika, 96(2), 371 - 382.

Inacio de Carvalho, V., and Rodriguez-Alvarez, M. X. (2018). Bayesian nonparametric inference for the covariate-adjusted ROC curve. arXiv preprint arXiv:1806.00473.

Rodriguez-Alvarez, M. X., Roca-Pardinas, J., and Cadarso-Suarez, C. (2011). ROC curve and covariates: extending induced methodology to the non-parametric framework. Statistics and Computing, 21(4), 483 - 499.


}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%}
%\examples{
%}
